add parentheses in FastAToCompact4BigBand to deconfuse gcc 4.1.2

+++ src/libmaus2/fastx/FastAToCompact4BigBand.hpp
@@ -191,7 +191,7 @@ namespace libmaus2
 							}
 						}
 
-						assert ( seqtmpOut->tellp() == static_cast<int64_t>(seqoff) );
+						assert ( static_cast<int64_t>(seqtmpOut->tellp()) == static_cast<int64_t>(seqoff) );
 						seqtmpOut->write(spattern.c_str(),seqlen);
 						seqoff += seqlen;
